Moved CustomizableToast to the main folder
