+++ Clients/src/presentation/pages/Authentication/RegisterMultiTenant/index.tsx
@@ -1,5 +1,5 @@
 import { Button, Stack, Typography, useTheme } from "@mui/material";
-import React, { Suspense, useState, useEffect} from "react";
+import React, { Suspense, useState, useEffect } from "react";
 import { ReactComponent as Background } from "../../../assets/imgs/background-grid.svg";
 import Check from "../../../components/Checks";
 import Field from "../../../components/Inputs/Field";
@@ -17,7 +17,7 @@ import type {
   OrganizationFormValues,
   OrganizationFormErrors,
 } from "../../../../application/validations/formValidation";
-import CustomizableToast from "../../../vw-v2-components/Toast";
+import CustomizableToast from "../../../components/Toast";
 import Alert from "../../../components/Alert";
 import { useDispatch } from "react-redux";
 import {
