+++ Clients/src/presentation/pages/SettingsPage/Password/index.tsx
@@ -15,7 +15,7 @@ import { extractUserToken } from "../../../../application/tools/extractToken";
 import CustomizableButton from "../../../components/Button/CustomizableButton";
 import SaveIcon from "@mui/icons-material/Save";
 import CustomizableSkeleton from "../../../components/Skeletons";
-import CustomizableToast from "../../../vw-v2-components/Toast"; // Import CustomizableToast
+import CustomizableToast from "../../../components/Toast"; // Import CustomizableToast
 import { updatePassword } from "../../../../application/repository/user.repository";
 
 const PasswordForm: React.FC = () => {
