+++ Clients/src/presentation/pages/Vendors/index.tsx
@@ -30,7 +30,7 @@ import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import AddNewRisk from "../../components/Modals/NewRisk";
 import CustomizableButton from "../../components/Button/CustomizableButton";
 import CustomizableSkeleton from "../../components/Skeletons";
-import CustomizableToast from "../../vw-v2-components/Toast";
+import CustomizableToast from "../../components/Toast";
 import RisksCard from "../../components/Cards/RisksCard";
 import { vwhomeHeading } from "../Home/1.0Home/style";
 import useVendorRisks from "../../../application/hooks/useVendorRisks";
